refactor(core): move files
